// src/features/parcours/dossiers-ds/domain/value-objects/ds-field-type.enum.ts
/**
 * Types de champs Démarches Simplifiées
 */
export enum DSFieldType {
  TEXT = "text",
  EMAIL = "email",
  PHONE = "phone",
  NUMBER = "integer_number",
  DATE = "date",
  CHECKBOX = "checkbox",
  FILE = "piece_justificative",
  TEXTAREA = "textarea",
  DROPDOWN = "drop_down_list",
  MULTIPLE_DROPDOWN = "multiple_drop_down_list",
  ADDRESS = "address",
  COMMUNE = "communes",
}